Size_t to avoid overflow

+++ jni/include/faiss_index_service.h
@@ -65,7 +65,7 @@ public:
     virtual void writeIndex(std::string indexPath, jlong idMapAddress);
     virtual ~IndexService() = default;
 protected:
-    virtual void allocIndex(faiss::Index * index, int dim, int numVectors);
+    virtual void allocIndex(faiss::Index * index, size_t dim, size_t numVectors);
     std::unique_ptr<FaissMethods> faissMethods;
 };
 
@@ -122,7 +122,7 @@ public:
     virtual void writeIndex(std::string indexPath, jlong idMapAddress) override;
     virtual ~BinaryIndexService() = default;
 protected:
-    virtual void allocIndex(faiss::Index * index, int dim, int numVectors) override;
+    virtual void allocIndex(faiss::Index * index, size_t dim, size_t numVectors) override;
 };
 
 }
